server/requirements.txt
Flask
numpy==1.23.5
pandas==1.5.3
tensorflow-cpu==2.14.0
requests
scikit-learn==1.2.2